backend/requirements.txt
fastapi
fastapi[standard]
pymongo
argon2-cffi
pyjwt
python-dotenv
uvicorn